splitEmojisはmisskey関係という訳でもないし大きいので個別ファイル化
