+++ src/components/GetNote.svelte
@@ -7,7 +7,8 @@
     emojis,
     defaultFFMpegArgs,
   } from "../lib/store";
-  import { getNote, splitEmojis, init } from "../lib/misskey";
+  import { getNote, init } from "../lib/misskey";
+  import { splitEmojis } from "../lib/splitEmojis";
   import { MiAuth } from "../lib/miauth.svelte";
 
   const miauth = new MiAuth();
